rename portability-model to portability-data-transfer-types
