+++ portability-data-transfer-types/src/main/java/org/dataportabilityproject/datatransfer/types/EntityType.java
@@ -1,4 +1,4 @@
-package org.dataportabilityproject.model;
+package org.dataportabilityproject.datatransfer.types;
 
 /**
  * A uniquely identifiable entity in the system.
