+++ portability-data-transfer-types/src/main/java/org/dataportabilityproject/datatransfer/types/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Defines the data portability model objects.  Model objects represent the flow of data through the system.
  */
-package org.dataportabilityproject.model;
+package org.dataportabilityproject.datatransfer.types;
