+++ portability-spi-cloud/build.gradle
@@ -1,5 +1,5 @@
 
 dependencies {
-    compile project(':portability-model')
+    compile project(':portability-data-transfer-types')
 
 }
