+++ settings.gradle
@@ -1,6 +1,6 @@
 rootProject.name = 'org.dataportabilityproject'
 
-include ':portability-model', ':portability-spi-cloud', ':portability-spi-data-transfer', ':portability-gateway', ':portability-worker-service'
+include ':portability-data-transfer-types', ':portability-spi-cloud', ':portability-spi-data-transfer', ':portability-gateway', ':portability-worker-service'
 
 include ':extensions:cloud:portability-cloud-google', ':extensions:cloud:portability-cloud-microsoft'
